requirements.txt
discord==2.0.0
discord.py==2.0.0
openai==0.27.4     
python-dotenv==0.20.0    
requests==2.28.2
